refactor: redesign mux-server DI — direct imports, grouped db/config
Replace injected pure functions with module-level imports, collapse
individual DB statement deps into typed `db: Pick<PreparedStatements, ...>`
objects, and group config values into `config: Pick<MuxConfig, ...>`.

server.ts reduced from 3487 to 848 lines. Factory dep counts drop from
30-76 to 8-24, with remaining deps being genuine DI (service handles,
mutable state, cross-cutting concerns).
